Added on deselect without breaking current working. Like it or not, firing onSelect with no selectected item (as per google charts originally) when deselecting a chart portion is not helpfull in any way shape or form. It does nothing for you but tel you that a selection has occured with no data available which is wrong, a selection was not made, a selection was released. I am not fussed wether this makes it in or not, but quoting current functionality as a basis for not adding this valuable event is silly, we are extending the charting with an invaluable tool, the ability to know which chart area was deselected.
The fact here is that google charts incorrectly fires a select event with no data when no selection has been made. onSelect should only fire when you select something.... Either google charts should not fire select when clicking the same chart area for the second time, or regardless of whether the chart area was previously selected or not, it should allow re-selection regardless and hence fire the selected event everytime.

I am fixing an issue with google charts here (without breaking current functionality), merge this or not, I have 4 instances where this requirement is a must, so I will be forced to support a fork should it not be merged..... or just rewrite the whole lot which I would rather not.